Introduce generating .pxd file
